sprint 14

+++ src/controllers/cards.js
@@ -2,20 +2,18 @@ const Card = require('../models/card');
 
 const Helpers = require('../utils/helpers');
 const AppErrors = require('../appErrors/appErrors');
+const BadRequestErr = require('../appErrors/bad-request-error');
+const NotFoundErr = require('../appErrors/not-found-err');
 
 const fields = 'name link owner likes';
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({}, fields).populate('owner', 'name about').populate('likes', 'name about')
-    // в ответе может быть owner: null,
-    // из-за захардкоженного _id пользователя в app.js, согласно заданию 13
     .then((result) => res.send({ data: result }))
-    .catch(() => res
-      .status(AppErrors.ERROR_CODE_SERVER_ERROR)
-      .send({ message: AppErrors.ERROR_SERVER }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((result) => {
@@ -27,108 +25,67 @@ module.exports.createCard = (req, res) => {
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res
-          .status(AppErrors.ERROR_CODE_BAD_REQUEST)
-          .send({ message: AppErrors.ERROR_NEW_CARD_PARAMS });
+        return next(new BadRequestErr(AppErrors.ERROR_NEW_CARD_PARAMS));
       }
-      return res
-        .status(AppErrors.ERROR_CODE_SERVER_ERROR)
-        .send({ message: AppErrors.ERROR_SERVER });
+      return next(err);
     });
 };
 
-module.exports.deleteCard = (req, res) => {
+module.exports.deleteCard = (req, res, next) => {
   const id = Helpers.getMongoId(req.params.cardId);
   if (!id) {
-    res.status(AppErrors.ERROR_CODE_BAD_REQUEST).send({ message: AppErrors.ERROR_PARAMS });
+    next(new BadRequestErr(AppErrors.ERROR_PARAMS));
     return;
   }
-
-  Card.findOneAndDelete({ _id: id })
+  // удаляем только карточку, которой мы владеем
+  Card.findOneAndDelete({ _id: id, owner: req.user._id })
     .then((result) => {
       if (!result) {
-        res
-          .status(AppErrors.ERROR_CODE_NOT_FOUND)
-          .send({ message: AppErrors.ERROR_CARD_NOT_FOUND });
-        return;
+        throw new NotFoundErr(AppErrors.ERROR_CARD_NOT_FOUND);
       }
       Card.find({}, fields).populate('owner', 'name about').populate('likes', 'name about')
-        // в ответе может быть owner: null,
-        // из-за захардкоженного _id пользователя в app.js, согласно заданию 13
         .then((cards) => res.send({ data: cards }))
-        .catch(() => res
-          .status(AppErrors.ERROR_CODE_SERVER_ERROR)
-          .send({ message: AppErrors.ERROR_SERVER }));
+        .catch(next);
     })
-    .catch(() => res
-      .status(AppErrors.ERROR_CODE_SERVER_ERROR)
-      .send({ message: AppErrors.ERROR_SERVER }));
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   const id = Helpers.getMongoId(req.params.cardId);
-  if (!id) {
-    res
-      .status(AppErrors.ERROR_CODE_BAD_REQUEST)
-      .send({ message: AppErrors.ERROR_PARAMS });
-    return;
-  }
-
+  if (!id) { next(new BadRequestErr(AppErrors.ERROR_PARAMS)); return; }
   Card.findByIdAndUpdate(
     id,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
     { new: true, fields },
   )
     .then((result) => {
-      if (!result) {
-        return res
-          .status(AppErrors.ERROR_CODE_NOT_FOUND)
-          .send({ message: AppErrors.ERROR_CARD_NOT_FOUND });
-      }
+      if (!result) throw new NotFoundErr(AppErrors.ERROR_CARD_NOT_FOUND);
       return res.send({ data: result });
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res
-          .status(AppErrors.ERROR_CODE_BAD_REQUEST)
-          .send({ message: AppErrors.ERROR_CARD_LIKES_PARAMS });
+        return next(new BadRequestErr(AppErrors.ERROR_CARD_LIKES_PARAMS));
       }
-      return res
-        .status(AppErrors.ERROR_CODE_SERVER_ERROR)
-        .send({ message: AppErrors.ERROR_SERVER });
+      return next(err);
     });
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   const id = Helpers.getMongoId(req.params.cardId);
-  if (!id) {
-    res
-      .status(AppErrors.ERROR_CODE_BAD_REQUEST)
-      .send({ message: AppErrors.ERROR_PARAMS });
-    return;
-  }
-
+  if (!id) { next(new BadRequestErr(AppErrors.ERROR_PARAMS)); return; }
   Card.findByIdAndUpdate(
     id,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
     { new: true, fields },
   )
     .then((result) => {
-      if (!result) {
-        return res
-          .status(AppErrors.ERROR_CODE_NOT_FOUND)
-          .send({ message: AppErrors.ERROR_CARD_NOT_FOUND });
-      }
+      if (!result) throw new NotFoundErr(AppErrors.ERROR_CARD_NOT_FOUND);
       return res.send({ data: result });
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res
-          .status(AppErrors.ERROR_CODE_BAD_REQUEST)
-          .send({ message: AppErrors.ERROR_CARD_LIKES_PARAMS });
+        return next(new BadRequestErr(AppErrors.ERROR_CARD_LIKES_PARAMS));
       }
-      return res
-        .status(AppErrors.ERROR_CODE_SERVER_ERROR)
-        .send({ message: AppErrors.ERROR_SERVER });
+      return next(err);
     });
 };
